chore: rename gameboard folder
