+++ game-app/src/app/page.tsx
@@ -1,6 +1,6 @@
 import Card from "./components/card/Card";
 import ConnectButton from "./components/ConnectButton";
-import Gameboard from "./components/Gameboard/Gameboard";
+import Gameboard from "./components/gameboard/Gameboard";
 import Scoreboard from "./components/Scoreboard";
 
 export default function Home() {
